requirements.txt
beautifulsoup4
tornado
pycrypto
MySQL-python
requests
lxml
sqlalchemy